Updated manuscript (main tex and pdf files)
Have merged (manually) the two different manuscript versions, and placed
as manuscript.tex, adding PDF and placing the figure files in the Fig
subfolder (renaming two of the files). Note that I do not now do the
LaTeX typesetting inside the Git folder, so the auxhiliary files are not
included including the tex-aux folder.